+++ packages/dialect-react-sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [UNRELEASED]
 
+- fix: update threads order after new message is sent
+
 ## [1.0.0-beta.2] - 2022-06-24
 
 - initial release with new react api implementation using sdk